+++ smol/cofe/wrangling/__init__.py
@@ -3,8 +3,8 @@
 And additional wrangling functions to preprocess and filter fitting data.
 """
 
-from .tools import filter_duplicate_corr_vectors, filter_by_ewald_energy, \
+from .tools import unique_corr_vector_indices, max_ewald_energy_indices, \
     weights_energy_above_composition, weights_energy_above_hull
 
-__all__ = ["filter_duplicate_corr_vectors", "filter_by_ewald_energy",
+__all__ = ["unique_corr_vector_indices", "max_ewald_energy_indices",
            "weights_energy_above_composition", "weights_energy_above_hull"]